Make members private and move a class.
